Initial run function implemented Fixed typing imports

+++ Individual.py
@@ -1,6 +1,6 @@
 
 
-from typing import List, Double, Int
+from typing import List
 from copy import deepcopy
 
 
@@ -8,7 +8,7 @@ class Individual():
 
 
     _representation : List[int] = list()
-    _distance: Double = 0
+    _distance: float = 0
 
     def __init__(self) -> None:
         pass
@@ -23,16 +23,16 @@ class Individual():
     def __gt__(self, value: object) -> bool:
         pass
 
-    def use_cyclic_notation(self, cyclic_notation: List[Int]) -> None:
+    def use_cyclic_notation(self, cyclic_notation: List[int]) -> None:
         pass
 
-    def use_path_notation(self, cyclic_notation: List[Int]) -> None:
+    def use_path_notation(self, cyclic_notation: List[int]) -> None:
         pass
 
     def get_cyclic_representation(self) -> List[int]:
         return deepcopy(self._representation)
     
-    def get_distance(self) -> Double:
+    def get_distance(self) -> float:
         #TODO: layzy implementation
         return self._distance
